+++ src/java.security.jgss/share/classes/sun/security/krb5/internal/crypto/dk/Des3DkCrypto.java
@@ -157,7 +157,7 @@ public class Des3DkCrypto extends DkCrypto {
             }
             ++posn;
             if (bit != 0) {
-                last |= (bit<<posn);
+                last |= (byte) (bit<<posn);
             }
         }
