+++ test/hotspot/jtreg/compiler/runtime/Test8010927.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2013, 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -27,8 +27,8 @@
  * @summary Kitchensink crashed with SIGSEGV, Problematic frame: v ~StubRoutines::checkcast_arraycopy
  * @library /test/lib
  * @modules java.base/jdk.internal.misc:+open
- * @build sun.hotspot.WhiteBox
- * @run driver jdk.test.lib.helpers.ClassFileInstaller sun.hotspot.WhiteBox
+ * @build jdk.test.whitebox.WhiteBox
+ * @run driver jdk.test.lib.helpers.ClassFileInstaller jdk.test.whitebox.WhiteBox
  * @run main/othervm -XX:+UnlockDiagnosticVMOptions -XX:+IgnoreUnrecognizedVMOptions
  *                   -XX:+WhiteBoxAPI -Xbootclasspath/a:. -Xmx128m -XX:NewSize=20971520
  *                   -XX:MaxNewSize=32m -XX:-UseTLAB -XX:-UseAdaptiveSizePolicy
@@ -38,7 +38,7 @@
 package compiler.runtime;
 
 import jdk.internal.misc.Unsafe;
-import sun.hotspot.WhiteBox;
+import jdk.test.whitebox.WhiteBox;
 
 import java.lang.reflect.Field;
 
@@ -74,7 +74,7 @@ public class Test8010927 {
     static final Test8010927 elem = new Test8010927();
     static final WhiteBox wb = WhiteBox.getWhiteBox();
 
-    static final int obj_header_size = U.ARRAY_OBJECT_BASE_OFFSET;
+    static final int obj_header_size = (int) U.ARRAY_OBJECT_BASE_OFFSET;
     static final int heap_oop_size = wb.getHeapOopSize();
     static final int card_size = 512;
     static final int one_card = (card_size - obj_header_size) / heap_oop_size;
